fix bugs

+++ src/PRO/l10/ex_004_lambdaCalculate/calculateWithLambdas/Main.java
@@ -5,7 +5,7 @@ public class Main {
         Operationable operation;
         operation = (x, y) -> x + y;
 
-        int result = operation.calculate(10, 20);
+        double result = operation.calculate(10, 20);
         System.out.println(result); //30
     }
 }
